+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT VERSION
 
 - chore: remove the use of `Object.values`
+- feat: add `getColumnManager` and `getDOMNode` methods
 
 ## v1.6.5 (2019-07-11)